Add mechanism to prevent note's subject update
- Add validator for domain/note
- Apply validator to domain/note

Because there was no mechanism to check the note update parameter,
every incoming update request was executed unchecked. The result
was that unintended update was executed with unintended consequences,
such as, updating of a note's subject.

The validator for domain/note is added and applied to check the parameter.
Any incoming update request which failed the validation will be ignored.
The mechanism ignore the incoming update request entirely instead of
ignoring part of it, because it is assumed that any update request with
illegitimate parameter is illegitimate.

The update only add validation case of requesting note's subject update.
Additional validator details would be needed to prevent other cases.